+++ test/unit/domain/providers/pub/pubSuggestionProvider.fixtures.ts
@@ -40,7 +40,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('version', '1.2.3', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('version', createTextRange(15, 15)),
+          createPackageNameDesc('version', createTextRange(15)),
           createPackageVersionDesc('1.2.3', createTextRange(24, 29)),
           createProjectVersionTypeDesc()
         ])
@@ -48,7 +48,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('efts', '^2.0.4', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('efts', createTextRange(376, 376)),
+          createPackageNameDesc('efts', createTextRange(376)),
           createPackageVersionDesc('^2.0.4', createTextRange(382, 388)),
           createPackageParentDescType('dependencies')
         ])
@@ -56,7 +56,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('http', '*', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('http', createTextRange(391, 391)),
+          createPackageNameDesc('http', createTextRange(391)),
           createPackageVersionDesc('*', createTextRange(397, 397), '', ' '),
           createPackageParentDescType('dependencies')
         ])
@@ -64,7 +64,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('transmogrify', '^0.4.0', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('transmogrify', createTextRange(421, 421)),
+          createPackageNameDesc('transmogrify', createTextRange(421)),
           createPackageVersionDesc('^0.4.0', createTextRange(448, 454), '', ''),
           createPackageParentDescType('dependencies')
         ])
@@ -72,7 +72,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('test', '>=0.5.0 <0.12.0', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('test', createTextRange(489, 489)),
+          createPackageNameDesc('test', createTextRange(489)),
           createPackageVersionDesc('>=0.5.0 <0.12.0', createTextRange(496, 511)),
           createPackageParentDescType('dependencies')
         ])
@@ -80,7 +80,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('collection', '^1.1.0', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('collection', createTextRange(515, 515)),
+          createPackageNameDesc('collection', createTextRange(515)),
           createPackageVersionDesc('^1.1.0', createTextRange(528, 534)),
           createPackageParentDescType('dependencies')
         ])
@@ -101,7 +101,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('pathify1', './some/test/path1', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('pathify1', createTextRange(17, 17)),
+          createPackageNameDesc('pathify1', createTextRange(17)),
           createPackagePathDescType('./some/test/path1', createTextRange(37, 54)),
           createPackageParentDescType('dependencies')
         ])
@@ -109,7 +109,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('pathify2', './some/test/path2', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('pathify2', createTextRange(57, 57)),
+          createPackageNameDesc('pathify2', createTextRange(57)),
           createPackagePathDescType('./some/test/path2', createTextRange(77, 94)),
           createPackageParentDescType('dependencies')
         ])
@@ -137,7 +137,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('gitify1', 'git@github.com:munificent/kittens.git', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('gitify1', createTextRange(17, 17)),
+          createPackageNameDesc('gitify1', createTextRange(17)),
           createPackageGitDescType('git@github.com:munificent/kittens.git'),
           createPackageParentDescType('dependencies')
         ])
@@ -145,7 +145,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('gitify2', 'git@github.com:munificent/dogs.git', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('gitify2', createTextRange(76, 76)),
+          createPackageNameDesc('gitify2', createTextRange(76)),
           createPackageGitDescType('git@github.com:munificent/dogs.git', '', 'some-branch'),
           createPackageParentDescType('dependencies')
         ])
@@ -153,7 +153,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('gitify3', 'git@github.com:munificent/birds.git', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('gitify3', createTextRange(167, 167)),
+          createPackageNameDesc('gitify3', createTextRange(167)),
           createPackageGitDescType('git@github.com:munificent/birds.git', 'path/to/birds'),
           createPackageParentDescType('dependencies')
         ])
@@ -161,7 +161,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('gitify4', 'git@github.com:munificent/foxes.git', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('gitify4', createTextRange(262, 262)),
+          createPackageNameDesc('gitify4', createTextRange(262)),
           createPackageGitDescType('git@github.com:munificent/foxes.git'),
           createPackageParentDescType('dependencies')
         ])
@@ -189,7 +189,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('hostify1', '1.0.0', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('hostify1', createTextRange(17, 17)),
+          createPackageNameDesc('hostify1', createTextRange(17)),
           createPackageVersionDesc('1.0.0', createTextRange(40, 45)),
           createPackageHostedDescType('https://some-package-server.com'),
           createPackageParentDescType('dependencies')
@@ -198,7 +198,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('hostify2', '2.0.0', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('hostify2', createTextRange(93, 93)),
+          createPackageNameDesc('hostify2', createTextRange(93)),
           createPackageVersionDesc('2.0.0', createTextRange(116, 121), '', ''),
           createPackageHostedDescType('https://some-package-server.com'),
           createPackageParentDescType('dependencies')
@@ -207,7 +207,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('hostify3', '3.0.0', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('hostify3', createTextRange(180, 180)),
+          createPackageNameDesc('hostify3', createTextRange(180)),
           createPackageVersionDesc('3.0.0', createTextRange(203, 208)),
           createPackageHostedDescType('https://some-package-server.com', 'testHostPackageAlias'),
           createPackageParentDescType('dependencies')
@@ -222,7 +222,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('version', '1.0.0', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('version', createTextRange(0, 0)),
+          createPackageNameDesc('version', createTextRange(0)),
           createPackageVersionDesc('1.0.0', createTextRange(9, 14)),
           createProjectVersionTypeDesc()
         ])
@@ -236,7 +236,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('version', '1.0.0', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('version', createTextRange(0, 0)),
+          createPackageNameDesc('version', createTextRange(0)),
           createPackageVersionDesc('1.0.0', createTextRange(10, 15)),
           createProjectVersionTypeDesc()
         ])
@@ -250,7 +250,7 @@ dependencies:
       new PackageDependency(
         createPackageResource('version', '1.0.0', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('version', createTextRange(0, 0)),
+          createPackageNameDesc('version', createTextRange(0)),
           createPackageVersionDesc('1.0.0', createTextRange(10, 15), '', ''),
           createProjectVersionTypeDesc()
         ])
@@ -264,9 +264,26 @@ dependencies:
       new PackageDependency(
         createPackageResource('version', '*', 'test/path'),
         new PackageDescriptor([
-          createPackageNameDesc('version', createTextRange(0, 0)),
-          createPackageVersionDesc('*', createTextRange(9, 9), '', ' '),
+          createPackageNameDesc('version', createTextRange(0)),
+          createPackageVersionDesc('*', createTextRange(9), '', ' '),
           createProjectVersionTypeDesc()
+        ])
+      )
+    ]
+  },
+
+  parsesAnyVersionKeyword: {
+    test: `
+    dependencies:
+      dep1: any
+    `,
+    expected: [
+      new PackageDependency(
+        createPackageResource('dep1', '*', 'test/path'),
+        new PackageDescriptor([
+          createPackageNameDesc('dep1', createTextRange(25)),
+          createPackageVersionDesc('*', createTextRange(31, 34)),
+          createPackageParentDescType('dependencies')
         ])
       )
     ]